Добавлена типо бд
