+++ app/src/main/AndroidManifest.xml
@@ -50,9 +50,6 @@ package="com.lapstr.lapstr">
         android:label="@string/title_activity_login"
         android:theme="@style/AppTheme.NoActionBar" />
     <activity
-        android:name=".DataBa"
-        android:theme="@style/AppTheme.NoActionBar" />
-    <activity
         android:name=".ResetPasswordActivity"
         android:label="@string/title_activity_reset_password"
         android:theme="@style/AppTheme.NoActionBar" />
