--- app/src/main/java/com/lapstr/lapstr/DataBa.java
@@ -1,136 +1,0 @@
-package com.lapstr.lapstr;
-
-/**
- * Created by Anatole on 25.01.2017.
- */
-
-import android.app.Activity;
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.text.TextUtils;
-import android.util.Log;
-import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.TextView;
-
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.ValueEventListener;
-
-public class DataBa extends Activity {
-
-    private static final String TAG = DataBa.class.getSimpleName();
-    private TextView txtDetails;
-    private EditText inputName, inputEmail;
-    private Button btnSave;
-    private DatabaseReference mFirebaseDatabase;
-    private FirebaseDatabase mFirebaseInstance;
-
-    private String userId;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.database_activity);//asfa
-
-        // Displaying toolbar icon
-    //    getSupportActionBar().setDisplayShowHomeEnabled(true);
-       // getSupportActionBar().setIcon(R.mipmap.ic_launcher);
-
-        txtDetails = (TextView) findViewById(R.id.txt_user);
-        inputName = (EditText) findViewById(R.id.name5);
-        inputEmail = (EditText) findViewById(R.id.email5);
-        btnSave = (Button) findViewById(R.id.btn_save5);
-
-        mFirebaseInstance = FirebaseDatabase.getInstance();
-
-        // get reference to 'users' node
-        mFirebaseDatabase = mFirebaseInstance.getReference("users");
-
-        // store app title to 'app_title' node
-        mFirebaseInstance.getReference("app_title").setValue("Realtime Database");
-
-        // app_title change listener
-        mFirebaseInstance.getReference("app_title").addValueEventListener(new ValueEventListener() {
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-                Log.e(TAG, "App title updated");
-
-                String appTitle = dataSnapshot.getValue(String.class);
-
-                // update toolbar title
-              //  getSupportActionBar().setTitle(appTitle);
-            }
-
-            @Override
-            public void onCancelled(DatabaseError error) {
-                // Failed to read value
-                Log.e(TAG, "Failed to read app title value.", error.toException());
-            }
-        });
-
-        // Save / update the user
-        btnSave.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                String name = inputName.getText().toString();
-                String email = inputEmail.getText().toString();
-                createUser(name, email);
-
-            }
-        });
-
-
-    }
-
-    /**
-     * Creating new user node under 'users'
-     */
-    private void createUser(String name, String email) {
-        // TODO
-        // In real apps this userId should be fetched
-        // by implementing firebase auth
-        userId = mFirebaseDatabase.push().getKey();
-        User user = new User(name, email);
-        mFirebaseDatabase.child(userId).setValue(user);
-        addUserChangeListener();
-    }
-
-    /**
-     * User data change listener
-     */
-    private void addUserChangeListener() {
-        // User data change listener
-        mFirebaseDatabase.child(userId).addValueEventListener(new ValueEventListener() {
-            @Override
-            public void onDataChange(DataSnapshot dataSnapshot) {
-                User user = dataSnapshot.getValue(User.class);
-
-                // Check for null
-                if (user == null) {
-                    Log.e(TAG, "User data is null!");
-                    return;
-                }
-
-                Log.e(TAG, "User data is changed!" + user.name + ", " + user.email);
-
-                // Display newly updated name and email
-                txtDetails.setText(user.name + ", " + user.email);
-
-                // clear edit text
-                inputEmail.setText("");
-                inputName.setText("");
-
-            }
-
-            @Override
-            public void onCancelled(DatabaseError error) {
-                // Failed to read value
-                Log.e(TAG, "Failed to read user", error.toException());
-            }
-        });
-    }
-}
