+++ app/src/main/java/com/lapstr/lapstr/User.java
@@ -4,19 +4,22 @@ package com.lapstr.lapstr;
  * Created by Anatole on 25.01.2017.
  */
 
+import android.net.Uri;
+
 import com.google.firebase.database.IgnoreExtraProperties;
+import com.google.firebase.storage.StorageReference;
 
 @IgnoreExtraProperties
 public class User {
 
     public String name;
-    public String email;
+    public String url;
 
     public User() {
     }
 
-    public User(String name, String email) {
+    public User(String name, String url) {
         this.name = name;
-        this.email = email;
+        this.url = url;
     }
 }
