+++ app/src/main/res/values/strings.xml
@@ -45,8 +45,4 @@
     <string name="btn_sign_out">Sign Out</string>
     <string name="lbl_forgot_password">Forgot password?</string>
     <string name="forgot_password_msg">We just need your registered Email Id to sent you password reset instructions.</string>
-
-    <string name="name5">name5</string>
-    <string name="email5">email5</string>
-    <string name="action_save5">action_save5</string>
 </resources>
